班長の掛け金設定を変更
